+++ KuchCraft/src/Renderer/Renderer.h
@@ -16,10 +16,10 @@ namespace KuchCraft {
 
 	};
 
-	constexpr size_t triangle_index_count = 6;
-	constexpr size_t quad_index_count     = 6;
-	constexpr size_t quad_vertex_count    = 4;
-	constexpr size_t cube_vertex_count    = 24;
+	constexpr uint32_t triangle_index_count = 6;
+	constexpr uint32_t quad_index_count     = 6;
+	constexpr uint32_t quad_vertex_count    = 4;
+	constexpr uint32_t cube_vertex_count    = 24;
 
 	class Chunk;
 	class Block;
